Do not use sendAction in the form components
