+++ app/components/form/closed-question.js
@@ -1,20 +1,34 @@
 import { ClosedQuestionTypes } from 'alpha-amber/constants';
+import { inject as service } from '@ember/service';
 import OpenQuestionComponent from './open-question';
 
 export default OpenQuestionComponent.extend({
+  store: service(),
   questionTypes: ClosedQuestionTypes,
   actions: {
     addOption() {
-      this.sendAction('onAddOption', this.question);
+      const position = this.question.get('sortedOptions.lastObject.position') + 1 || 0;
+      this.store.createRecord('form/closed-question-option', {
+        question: this.question,
+        position
+      });
     },
     deleteOption(option) {
-      this.sendAction('onDeleteOption', option);
+      option.deleteRecord();
     },
     moveOptionUp(option) {
-      this.sendAction('onMoveOptionUp', option, this.question);
+      const index = this.question.get('sortedOptions').indexOf(option);
+      if (index > 0) {
+        const previousOption = this.question.get('sortedOptions').objectAt(index - 1);
+        this.send('switchPositions', option, previousOption);
+      }
     },
     moveOptionDown(option) {
-      this.sendAction('onMoveOptionDown', option, this.question);
+      const index = this.question.get('sortedOptions').indexOf(option);
+      if (index < (this.question.get('sortedOptions.length') - 1)) {
+        const nextOption = this.question.get('sortedOptions').objectAt(index + 1);
+        this.send('switchPositions', option, nextOption);
+      }
     }
   }
 });
